Get rid of 'cooked' timings
As a convenience, timeall was computing 'cooked' (startup subtracted)
timings for each test and saving them to the timings files.  But
`timeall` shouldn't be cooking anything -- it's just a data gathering
program.  It's also not terribly useful since `analyze` has to handle
subtracting compile time for scalable tests anyway.  So just save the
compiler startup times in the 'run' section of the timings file, and
get rid of the 'cooked' computations and all the cruft that went with
them.