Add rc-forest-fire-stringify mini-benchmark
This is the second benchmark inspired by the slowness of rc-forest-fire,
in this case testing the stringification of an rc-forest-fire map from
an AoA of cell states to a single string ready for output.  This
stresses array slice with repeated keys, join with many short elements,
and either push or gather/take depending on the language.