Add create_and_copy_2d_grid_cross test; emulate cross for perl5
This is the third benchmark inspired by the slowness of rc-forest-fire,
in this case testing creation and copying of 2D grids using for/cross
to iterate over the grids.

Because Perl 5 has no real iterator-style cross operator, I also add a
cross-of-ranges emulation based on an iterating closure and use it for
the perl5 variant of this test and visit_2d_indices_cross.