Skip noise timings when analyzing
Timing data that is barely above the ignore values (startup and
compile times) is essentially noise and leads to junk values on
the left side of plots, and more annoyingly, screws up peak rates
used in many places.  Just skip these for the greater good.

This has the annoying side effect of making enough of the left
side of plots for very fast compilers disappear that there may
end up being no vertical alignment at all of the plots of fast
and slow compilers, but the trends generally remain clear.