Don't need to limit test tags in `bench quickstart`
This is no longer needed now that --skip-incomplete exists and is default.
It could save some testing time if the user only ever wanted to compare
against NQP, but it seems more useful to default to testing everything
that can be tested for each compiler, and let `analyze` decide what data
to use.